Complete Resource Struct
The resource struct now includes values for request, allocation, and release as well as arrays for each child process index to track individual allocation; reqArray, allArray, and relArray.
The first 3-5 resources (randomly determined: 15-25%) are initialized to 1, indicating that they are shared. The remaining resources are initialized to 0, indicating that they are not shared. Shared resources are not considered in the Deadlock Detection Algorithm (DDA).